Move code wrap into language-specific methods file.
Previously, the code wrap function only used spaces to indent and did
not wrap perl comments at all.  Now, we can specify on a language by
language basis what lines we want to match and how to indent wrapped
versions of those lines. Right now, it means we can switch to indenting
perl with tabs and begin to wrap perl's single-line comments.  In the
future, we will be able to be much more flexible in how we wrap code.